Fix object id handling bugs on 64-bit systems
A Ruby object_id is an unsigned long.  The inspect method for
GSL::Matrix::Complex and GSL::Vector::Complex include the object's
object_id in the returned String.  The code that did this converted the
object_id VALUE to an integer form using NUM2UINT.  That works OK on
32-bit systems, but leads to errors on 64-bit systems.  The fix is to
change NUM2UINT to NUM2ULONG.

git-svn-id: svn+ssh://rubyforge.org/var/svn/rb-gsl/trunk/rb-gsl@121 6e764f74-f39f-46f8-8c54-8307d62afe8d