Hack rdtordoc.rb for one-time conversion of rb-gsl
This includes several hacks to facilitate the onetime conversion of rd
docs to rdoc docs.  It includes things like hacking in anchor names to
rdoc headings and simplifying irb prompts.

Manual tweaking will need to be done on the generated rdoc files, but I
think this script is to the point where time will be better spent
tweaking the generated rdocs than tweaking this script to handled
everything perfectly.

git-svn-id: svn+ssh://rubyforge.org/var/svn/rb-gsl/trunk/rb-gsl@125 6e764f74-f39f-46f8-8c54-8307d62afe8d