It's a monstrosity. It compiles, doesn't run.
On the way to our NMatrix type! Dense is mostly implemented. List storage needs to be connected too.

Most likely there are some regular bugs that need to be worked out.

Additionally, still need to figure out how mark-and-sweep works. Using malloc in a lot of places where I think I need to be using the Ruby ALLOC_N macro.