Beginnings of matrix multiplication using BLAS and ATLAS.
Note that this version has a segfault problem. Even if that worked, it would only support double precision matrices.

In the spec, "dense correctly handles double multiplication", it'll work just fine if we change m's dimensions to 3x3 -- may have to fiddle with the major stride of the result in cblas_dgemm call too. For whatever reason, this is causing a major malloc problem.

With that said, linking works (kind of) correctly, which is a big deal. Took all day, you might say.