Added NVector. For this, was necessary to slightly refactor nm_alloc and nm_init. Should make NMatrix more consistent with other Ruby extensions.
Added vector multiplication for dense (gemv), which may or may not work consistently.

Made parameters consistent with DENSE_PARAM for function pointers. Yale will use YALE_PARAM, but will function slightly differently; DENSE_PARAM is for matrices A, B, and C in a gemm operation but YALE_PARAM is for a single matrix's ia, ja, and a pointers.