All matrix conversions (stype and dtype) now work.
One word of caution: when we chain a whole bunch of scasts together, it seems to cause a garbage collector loop of some kind. It's hard to debug, unfortunately, because it works fine when I use rake spec:gdb and rake spec:valgrind, but segfaults when I use irb (rake console). I haven't been able to figure out how to attach gdb or valgrind to IRB.