Limit files included in gem
The gemspec used to inlcude all files matching "ext/*" in the gem file.
This caused problems when extconf.rb and make had been run in the ext
subdirectory because in included files that are supposed to be generated
when the gem is being installed (e.g. Makefile and object files).  Now
the gemspec explicitly only inlcudes "ext/extconf.rb" and "ext/*.c".