Allow use of --with-cflags during gem install
extconf.rb now appends to $CFLAGS rather than ignoring its initial
value.  This allows additional CFLAGS to be given on the gem install
line.  For example, to build Ruby/GSL with debugging info in the native
extentions:

gem install gsl -- --with-cflags=-g

NB: The extra " -- " in the command line is not a typo.