Add test to expose bug that crashes Ruby process
This test exposes a bug that manifests itself when Data_Get_Vector is
passed an NArray object rather than a GSL::Vector object.  One method
where this condition can be excited is GSL::Blas::dnrm2.  Other methods
might expose this bug as well, but this test uses dnrm2.  This test can
be used to verify the forthcoming fix for this bug.