Fix bug when certain methods are passed NArray
Many methods can be passed an NArray or a GSL::Vector.  Some of these
methods use the Data_Get_Vector macro to get a pointer to a gsl_vector
that corresponds to the NArray or GSL::Vector argument.  When
Data_Get_Vector was given an NArray, it assumed that the given
gsl_vector pointer pointed to a valid gsl_vector structure, but this is
not at all a valid assumption.  In fact, the gsl_vector pointer given is
to be initialized by the Data_Get_Vector (its current value is ignored).

The fix is to convert the given NArray object to a GSL::Vector::View
object and initialize the gsl_vector pointer from the
GSL::Vector::View's gsl_vector_view pointer.  To do this, the
rb_gsl_na_to_gsl_vector_view_method function in gsl_narray.c is made
non-static and called from within Data_Get_Vector when an NArray object
is given.