Fix unused and set-but-not-used warnings
GCC 4.6 warns about unused local variables and "set but not used" local
variables.  This commit addresses these warnings when their effect is
benign.  Some of these warnings actually revealed bugs, but those bugs
will be fixed in later commits (after tests are added to expose them).

Many of the warnings addressed by this commit involve the local variable
"status" which is set to the return value of the GSL function that is
being called, but then never checked.  These cases should ideally check
the return value and raise an exception on any returned error, but for
now the "status" variable has been commented out (to facilitate
addinf these checks in the future) rather than being deleted (as was
done for other set-but-unused variables).