Add matrix_complex_test.rb
This test script exposes a bug in:

GSL::Matrix::Complex.eye(Integer, GSL::Complex)

The GSL::Complex value passed in is ignored and an indeterminate value
is used in its place.

Note that this test script uses minitest, which is part of the Ruby 1.9
standard library and available for Ruby 1.8 as a gem.